added may warning flags to CMakeLists

+++ src/util/pgn.cpp
@@ -10,10 +10,10 @@ pgn::pgn(const std::string& file_name)
     file.getline(line,line_size);
     if(line[0] == '['){
       std::string line_str(line);
-      int key_end = line_str.find_first_of(' ',0);
+      size_t key_end = line_str.find_first_of(' ',0);
       std::string key = line_str.substr(1,key_end-1);
-      int value_start = line_str.find_first_of('\"',key_end-1) + 1;
-      int value_end = line_str.find_first_of('\"',value_start+1);
+      size_t value_start = line_str.find_first_of('\"',key_end-1) + 1;
+      size_t value_end = line_str.find_first_of('\"',value_start+1);
       std::string value = line_str.substr(value_start,value_end-value_start);
       data[key] = value;
     }
